[SWM-169] Feat : develop summary generator ver2
