+++ app/gpt/gpt.py
@@ -20,43 +20,3 @@ def request_gpt(prompt: str):
     )
 
     return response["choices"][0]["message"]["content"]
-
-
-def generate_summary(script: Script):
-    summary_prompt = constants['prompt']['summary']
-    prompt = script.text + summary_prompt
-
-    summary = request_gpt(prompt)
-
-    return summary
-
-
-def reformat_summary(summary: str):
-    summary = summary.replace("\n", "\\n")
-    summary = summary.replace("\"", "\\\"")
-
-    return summary
-
-
-def generate_quiz(summary: str):
-    quiz_prompt = constants['prompt']['quiz']
-
-    prompt = summary + quiz_prompt
-
-    quiz_json = ''
-    quiz_json += request_gpt(prompt)
-
-    quiz_json = reformat_quiz(quiz_json)
-
-    return quiz_json
-
-
-def reformat_quiz(quiz_json: str):
-    quiz_json = quiz_json.replace("\n", "")
-    quiz_json = quiz_json.replace("\"", '"')
-
-    quiz_json = quiz_json[1:-1]
-
-    return quiz_json
-
-
