+++ constants.yaml
@@ -20,7 +20,7 @@ model_parameter :
 
 prompt :
   system_message : {"role": "system", "content": "You are an assistant that generates quizzes and summaries"}
-  summary : "\n\n 위 스크립트를 최대한 모든 내용이 반영될 수 있도록 요약해줘"
+  summary : "\n\n 위 스크립트를 최대한 모든 내용이 반영될 수 있도록 요약해줘. \"위 스크립트는\" 과 같은 서론은 빼줘"
   quiz : "\n\n 위 요약본을 바탕으로 객관식 1문제, 주관식 1문제, True or False 1문제 내줘. \n
           JSON 형식으로 답변해주고 아래 형식을 지켜셔 답변해줘. \n
           quizes: [{\"quizType\":\"\" , \"quizQuestion\":\"\", \"quizSelectOption1\": \"\", \"quizSelectOption2\": \"\", \"quizSelectOption3\": \"\", \"quizSelectOption4\": \"\", \"answer\":\"\"}] \n
